chore: 패키지 구조 변경으로 인한 충돌로 다시 기존 상태로 변경
