+++ src/main/java/com/example/freshcart/domain/Role.java
@@ -1,0 +1,8 @@
+package com.example.freshcart.domain;
+
+/**
+ * 가입 유형: 유저, 판매자, 관리자
+ */
+public enum Role {
+    USER, SELLER, ADMIN
+}
