+++ src/main/java/com/example/freshcart/presentation/web/argumentresolver/LoginCheck.java
@@ -1,0 +1,16 @@
+package com.example.freshcart.presentation.web.argumentresolver;
+
+/**
+ * 해당 경로에는 사용자 인증이 강제됨.
+ */
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginCheck {
+
+}
