+++ src/main/java/com/example/freshcart/product/domain/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.example.freshcart.product.domain;
 
-import com.example.freshcart.user.domain.User;
+import com.example.freshcart.domain.User;
 import java.util.List;
 import java.util.Optional;
