--- src/main/java/com/example/freshcart/user/domain/Role.java
@@ -1,8 +1,0 @@
-package com.example.freshcart.user.domain;
-
-/**
- * 가입 유형: 유저, 판매자, 관리자
- */
-public enum Role {
-    USER, SELLER, ADMIN
-}
