--- src/main/java/com/example/freshcart/user/infrastructure/exception/PasswordDoesNotMatchException.java
@@ -1,22 +1,0 @@
-package com.example.freshcart.user.infrastructure.exception;
-
-import org.springframework.http.HttpStatus;
-
-/**
- * 로그인 시 이메일 존재 여부/패스워드 체크
- */
-
-public class PasswordDoesNotMatchException extends BaseException {
-
-  private static final String message = "아이디 또는 비밀번호가 일치하지 않습니다";
-
-  public PasswordDoesNotMatchException() {
-    super(HttpStatus.BAD_REQUEST, message);
-  }
-
-  @Override
-  public String getMessage() {
-    return message;
-  }
-
-}
